requirements.txt
alembic==1.7.7
cffi==1.15.0
click==8.1.3
cryptography==37.0.1
ecdsa==0.17.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
install==1.3.5
itsdangerous==2.1.2
Jinja2==2.10.1
jwt==1.3.1
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycparser==2.21
PyJWT==1.7.1
python-jose==3.3.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
zipp==3.8.0
